x64: jit_generator: narrow small-immediate values per review
Per densamoilov's review on #5565 (brgemm_containers.cpp:52), remove the
dim_t overloads for instructions that only ever receive small, fixed-range
immediates: pextrb, pinsrb, vpextrb, vpinsrb, vpermq, vpblendd, vcmpps, and
shl. None of these overloads had any call site passing a dim_t value
(verified across all of src/cpu/x64); the two shl call sites in
jit_brgemm_kernel.cpp relied on it only because typesize_C/typesize_D were
widened to dim_t upstream, so fix those with an explicit, provably-bounded
cast instead (type sizes are always 1/2/4/8).

Keep the add/sub/cmp/imul dim_t overloads: imul is genuinely used to scale
a runtime stride into a byte offset (jit_avx512_core_resampling.cpp and
others), so it belongs with add/sub/cmp as a real tensor-scale boundary
case, not a small fixed-value one.

Also narrow jit_brdgmm_kernel_base_t::simd_w_ and max_vmms_ back to int:
they come from vlen/typesize and isa_num_vregs(), both hardware/ISA-bounded
quantities, not tensor-scale values. This also fixes a latent narrowing bug
in get_substep_simd() where the else-branch returned a dim_t simd_w_ from
an int-returning function.

Verified via targeted -Wshorten-64-to-32 syntax checks: zero new warnings
across the 131 files in src/cpu/x64 that call any of the touched
instructions, compared against the pre-change baseline (only the known
pre-existing xbyak_util.h:1594 warning remains).

+++ src/cpu/x64/brgemm/jit_brgemm_kernel.cpp
@@ -1132,13 +1132,19 @@ void jit_brgemm_kernel_t<Wmm>::read_params() {
 
     if (brg.is_runtime_ldc) {
         mov(reg_stride_ld_block, ptr[param1 + GET_OFF(dynamic_LDC)]);
-        if (brg.typesize_C > 1) shl(reg_stride_ld_block, (brg.typesize_C >> 1));
+        // typesize_C is always 1/2/4/8, so this shift amount is provably
+        // bounded regardless of typesize_C's declared (dim_t) type.
+        if (brg.typesize_C > 1)
+            shl(reg_stride_ld_block, static_cast<int>(brg.typesize_C >> 1));
         reg_stride_ld_block.save();
     }
 
     if (brg.is_runtime_ldd) {
         mov(reg_D_shift_bytes, ptr[param1 + GET_OFF(dynamic_LDD)]);
-        if (brg.typesize_D > 1) shl(reg_D_shift_bytes, (brg.typesize_D >> 1));
+        // typesize_D is always 1/2/4/8, so this shift amount is provably
+        // bounded regardless of typesize_D's declared (dim_t) type.
+        if (brg.typesize_D > 1)
+            shl(reg_D_shift_bytes, static_cast<int>(brg.typesize_D >> 1));
         reg_D_shift_bytes.save();
     }
